refactor: move classmap to its dedicated component
This the first step toward supporting exclusion on the frontend.
