+++ components/.ordering
@@ -28,6 +28,7 @@ configuration-accessor
 configuration-process
 configuration-environment
 location
+classmap
 trace
 backend
 mapping
