+++ components/classmap/default/.env
@@ -1,0 +1,2 @@
+test
+node
