+++ components/classmap/default/entity.test.mjs
@@ -1,4 +1,4 @@
-import { assertEqual, assertDeepEqual } from "../../../__fixture__.mjs";
+import { assertEqual, assertDeepEqual } from "../../__fixture__.mjs";
 
 import { parseEstree } from "./parse.mjs";
