+++ components/classmap/default/parse.test.mjs
@@ -1,4 +1,4 @@
-import { assertDeepEqual } from "../../../__fixture__.mjs";
+import { assertDeepEqual } from "../../__fixture__.mjs";
 import {
   parseEstree,
   printComment,
